+++ app/or_models_list.py
@@ -1,3 +1,4 @@
 models_list = {
-    1 : { "api_name" : "minimax/minimax-m2:free", "pretty_name" : "Minimax M2"}
+    1 : { "api_name" : "minimax/minimax-m2:free", "pretty_name" : "Minimax M2"},
+    2 : { "api_name" : "test/link-to-nowhere", "pretty_name" : "ChatGPT 5.2"}
 }